Added Homescreen and other Functionalities

+++ src/service/Adminservice.js
@@ -1,6 +1,8 @@
 import axios from "axios";
 
 const url = "http://localhost:3000/ticketlist";
+const url2 = "http://localhost:3000/admins";
+const url3 = "http://localhost:3000/adminspwd";
 
 const getAll = () => {
   return axios.get(url);
@@ -10,8 +12,24 @@ const getById = (id) => {
   return axios.get(url + `/${id}`);
 };
 
+const getAdminById = (id) => {
+  return axios.get(url2 + `/${id}`);
+};
+
+const getAdmin = () => {
+  return axios.get(url2);
+};
+
 const deleteById = (id) => {
   return axios.delete(url + `/${id}`);
 };
 
-export default { getAll, getById, deleteById };
+const changePassword = (id,data) => {
+  return axios.put(url3 + `/${id}`, data);
+};
+
+const updateStatus = (id,data) => {
+  return axios.put(url+ `/${id}`,data);
+};
+
+export default { getAll, getById, deleteById, getAdmin, getAdminById, changePassword, updateStatus };
